add navigation from main screen to uploadImage screen
